:recycle: (PassPlaylist) Extract SeeMoreWithEye
